refactor backend to static summaries
