+++ src/artifactminer/RepositoryIntelligence/repo_intelligence_user.py
@@ -10,10 +10,14 @@ from sqlalchemy import inspect, or_
 from artifactminer.db.database import SessionLocal
 from artifactminer.RepositoryIntelligence.repo_intelligence_main import isGitRepo, Pathish
 from artifactminer.RepositoryIntelligence.activity_classifier import classify_commit_activities 
-from artifactminer.RepositoryIntelligence.repo_intelligence_AI import user_allows_llm, createSummaryFromUserAdditions, saveUserIntelligenceSummary, group_additions_into_blocks
 from email_validator import validate_email, EmailNotValidError
 from sqlalchemy.orm import Session
-from artifactminer.db.models import RepoStat, UserRepoStat, UserAnswer
+from artifactminer.db.models import (
+    RepoStat,
+    UserRepoStat,
+    UserAnswer,
+    UserAIntelligenceSummary,
+)
 from pathlib import Path as PathLib
 
 @dataclass
@@ -207,15 +211,28 @@ def saveUserRepoStats(stats: UserRepoStats, db=None):
         if own_session:
             db.close()
 
+
+def saveUserIntelligenceSummary(repo_path: str, user_email: str, summary_text: str):
+    db = SessionLocal()
+    try:
+        user_summary = UserAIntelligenceSummary(
+            repo_path=repo_path,
+            user_email=user_email,
+            summary_text=summary_text,
+        )
+        db.add(user_summary)
+        db.commit()
+        db.refresh(user_summary)
+    finally:
+        db.close()
+
 async def generate_summaries_for_ranked(db: Session, top=3, extraction_path: str = None) -> list[dict]:
     """
     Summarize the top-ranked repositories for the current user.
 
     Expected flow (called by Nathan's orchestrator):
     - Shlok's ranking code has already set RepoStat.ranking_score.
-    - We pick the top 3, check consent, and either:
-      * use LLM summaries from diffs, or
-      * fall back to a simple template.
+    - We pick the top 3 and generate lightweight static summaries.
     - We persist results into UserAIntelligenceSummary.
     
     Note: Processes all repos concurrently for maximum performance.
@@ -289,28 +306,6 @@ async def generate_summaries_for_ranked(db: Session, top=3, extraction_path: str
                 f"User served as {role_text} on {repo.project_name} "
                 f"and contributed {pct:.1f}% using {languages_text}."
             )
-
-        # If we have consent + a valid email, try the LLM-based summary instead
-        if user_allows_llm() and user_email:
-            try:
-                # Ensure path is absolute for git operations
-                repo_path_absolute = str(Path(repo.project_path).resolve())
-                additions = collect_user_additions(
-                    repo_path=repo_path_absolute,
-                    user_email=user_email,
-                )
-                if additions:
-                    grouped = group_additions_into_blocks(additions, max_chars_per_block=1000, max_blocks=1)
-                    ai_summary = await createSummaryFromUserAdditions(grouped)
-                    print("\n AI Summary Generated for", repo.project_name, ":", ai_summary)
-                    summary_text += " AI summary: " + ai_summary
-            except Exception as e:
-                # Log the actual error for debugging
-                print(f"ERROR generating AI summary for {repo.project_name}: {type(e).__name__}: {e}")
-                import traceback
-                traceback.print_exc()
-                # Fail soft: keep the fallback template summary
-                pass
 
         # Persist into UserAIntelligenceSummary
         saveUserIntelligenceSummary(
